+++ example_runner.py
@@ -16,6 +16,10 @@ print('Window Examples')
 import example.tower.frame_window
 import example.tower.lattice_window
 
+print('Tile Examples')
+import example.tower.tile_generator
+import example.tower.tile_generator_dwarf_star
+
 print('Tower Examples')
 print('start tower_example')
 t1 = time.perf_counter(), time.process_time()