+++ changes.md
@@ -1,7 +1,12 @@
 ## Main Wip
+## 0.1.3
+* Upped cqterrain version 1.2.2
+* Added tile_generator example
+* Added tile_generator_dwarf_star example
+
 ## 0.1.2
 * Added magnets utility
-* TowerBase, TowerMid, ToerTop create magnet cutouts.
+* TowerBase, TowerMid, TowerTop create magnet cutouts.
 * Added TileGenerator proof of concept.
 
 ## 0.1.1